requirements.txt
langchain
langchain-community
langchain-openai
openai
streamlit
python-dotenv